Move LOOLWebSocket into test/ - where it can be slowly removed.
